requirements.txt
flask==2.3.3
gunicorn==21.2.0
numpy==1.26.4
pandas==2.1.4
scikit-learn==1.3.2
h5py==3.10.0
tensorflow-cpu==2.15.0
matplotlib
plotly
yfinance